requirements.txt
aioredis==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
azure-core==1.24.2
azure-storage-blob==12.13.1
bcc==0.18.0
bidict==0.22.1
billiard==3.6.4.0
blinker==1.7.0
Brlapi==0.8.3
celery==5.1.2
certifi==2020.6.20
cffi==1.16.0
channels==3.0.5
channels-redis==3.4.0
chardet==4.0.0
charset-normalizer==2.0.12
chrome-gnome-shell==0.0.0
colorama==0.4.4
command-not-found==0.3
constantly==23.10.4
cryptography==40.0.2
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.7
distro==1.7.0
Django==3.2.20
django-celery-beat==2.2.1
django-cors-headers==3.10.1
django-filter==21.1
django-timezone-field==4.2.3
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-jwt==1.19.2
et-xmlfile==1.1.0
filelock==3.12.4
Flask==3.0.0
Flask-SocketIO==5.3.6
gyp==0.1
h11==0.14.0
hidpidaemon==18.4.6
hiredis==2.3.2
httplib2==0.20.2
hyperlink==21.0.0
icalendar==5.0.11
idna==3.3
importlib-metadata==4.6.4
incremental==22.10.0
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
kernelstub==3.1.4
keyring==23.5.0
kombu==5.3.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==2.1.3
more-itertools==8.10.0
msgpack==1.0.7
msrest==0.7.1
mysql-connector==2.2.9
netaddr==0.8.0
netifaces==0.11.0
numpy==1.26.3
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
passlib==1.7.4
Pillow==8.4.0
platformdirs==3.11.0
plotly==5.18.0
pop-transition==1.1.2
prompt-toolkit==3.0.43
protobuf==3.12.4
psycopg2-binary==2.9.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.42.1
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.1.1
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-crontab==3.0.0
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-engineio==4.8.0
python-gnupg==0.4.8
python-socketio==5.10.0
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==4.3.6
repolib==2.2.1
repoman==1.4.0
requests==2.27.1
requests-oauthlib==1.3.1
SecretStorage==3.3.1
service-identity==24.1.0
sessioninstaller==0.0.0
simple-websocket==1.0.0
six==1.16.0
sqlparse==0.4.4
stripe==7.12.0
systemd-python==234
tenacity==8.2.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.4
ua-parser==0.18.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
user-agents==2.2.0
vine==5.1.0
virtualenv==20.24.5
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.0.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==6.1
